# services/aetherus-orbital/openapi/aetherus-v2.openapi.yaml
openapi: 3.1.0
info:
  title: Aetherus V2 API
  description: SPACE + CONTROL + ORBIT + INTELLIGENCE local product runtime. Scientific validation states are explicit.
  version: 0.5.0
paths:
...
  /v1/my-aetherus:
    get:
      summary: My Aetherus
      operationId: my_aetherus_v1_my_aetherus_get
      parameters:
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/my-aetherus/context:
    put:
      summary: Update Context
      operationId: update_context_v1_my_aetherus_context_put
      parameters:
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalContextUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/control/workspace:
    get:
      summary: Control Workspace
      operationId: control_workspace_v1_control_workspace_get
      parameters:
      - name: mission_state
        in: query
        required: false
        schema:
          type: string
          default: PRE-LAUNCH
          title: Mission State
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Save Workspace
      operationId: save_workspace_v1_control_workspace_put
      parameters:
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/follows/{target_id}:
    post:
      summary: Follow Target
      operationId: follow_target_v1_follows__target_id__post
      parameters:
      - name: target_id
        in: path
        required: true
        schema:
          type: string
          title: Target Id
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/follows/{target_id}/alerts:
    get:
      summary: Follow Alerts
      operationId: follow_alerts_v1_follows__target_id__alerts_get
      parameters:
      - name: target_id
        in: path
        required: true
        schema:
          type: string
          title: Target Id
      - name: revision_no
        in: query
        required: true
        schema:
          type: integer
          minimum: 1
          title: Revision No
      - name: summary
        in: query
        required: false
        schema:
          type: string
          default: Revision changed
          title: Summary
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/assistant/tool:
    post:
      summary: Assistant Tool
      operationId: assistant_tool_v1_assistant_tool_post
      parameters:
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantToolRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/research/datasets:
    get:
      summary: List Research Datasets
      operationId: list_research_datasets_v1_research_datasets_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      summary: Create Research Dataset
      operationId: create_research_dataset_v1_research_datasets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchDatasetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/research/datasets/{dataset_key}:
    get:
      summary: Download Research Dataset
      operationId: download_research_dataset_v1_research_datasets__dataset_key__get
      parameters:
      - name: dataset_key
        in: path
        required: true
        schema:
          type: string
          title: Dataset Key
      - name: format
        in: query
        required: false
        schema:
          type: string
          pattern: ^(manifest|json|csv)$
          default: manifest
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/operations/private/{key}:
    put:
      summary: Put Private State
      operationId: put_private_state_v1_operations_private__key__put
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivateStateUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Get Private State
      operationId: get_private_state_v1_operations_private__key__get
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/operations/audit:
    get:
      summary: Operations Audit
      operationId: operations_audit_v1_operations_audit_get
      parameters:
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/operations/jobs:
    post:
      summary: Submit Operation Job
      operationId: submit_operation_job_v1_operations_jobs_post
      parameters:
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationJobRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/operations/jobs/{job_id}:
    get:
      summary: Operation Job
      operationId: operation_job_v1_operations_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: x-aetherus-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-Tenant
      - name: x-aetherus-user
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Aetherus-User
      - name: x-aetherus-plan
        in: header
        required: false
        schema:
          type: string
          default: FREE
          title: X-Aetherus-Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AssistantToolRequest:
      properties:
        tool_name:
          type: string
          title: Tool Name
        args:
          type: object
          title: Args
        allow_scientific_tool:
          type: boolean
          title: Allow Scientific Tool
          default: false
      type: object
      required:
      - tool_name
      title: AssistantToolRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OperationJobRequest:
      properties:
        operation:
          type: string
          title: Operation
        payload:
          type: object
          title: Payload
        idempotency_key:
          type: string
          title: Idempotency Key
        run_now:
          type: boolean
          title: Run Now
          default: true
        max_attempts:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Max Attempts
          default: 3
      type: object
      required:
      - operation
      - idempotency_key
      title: OperationJobRequest
    PersonalContextUpdate:
      properties:
        context:
          type: object
          title: Context
      type: object
      title: PersonalContextUpdate
...
    PrivateStateUpdate:
      properties:
        value:
          title: Value
      type: object
      required:
      - value
      title: PrivateStateUpdate
...
    ResearchDatasetRequest:
      properties:
        dataset_key:
          type: string
          title: Dataset Key
        version:
          type: string
          title: Version
        domain:
          type: string
          title: Domain
        record_type:
          type: string
          title: Record Type
        license_policy:
          type: string
          title: License Policy
      type: object
      required:
      - dataset_key
      - version
      - domain
      - record_type
      - license_policy
      title: ResearchDatasetRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkspaceUpdate:
      properties:
        workspace:
          type: object
          title: Workspace
      type: object
      title: WorkspaceUpdate